+++ src/main/scala/tscfg/Main.scala
@@ -37,7 +37,6 @@ object Main {
        |  --java:optionals      use optionals                    (false)
        |  --scala               generate scala code              (java)
        |  --scala:bt            use backticks (see #30)          (false)
-       |  --scala:fp            report full path (see #36)       (false)
        |  --durations           use java.time.Duration           (false)
        |  --all-required        assume all properties are required (see #47)
        |  --tpl <filename>      generate config template         (no default)
@@ -55,7 +54,6 @@ object Main {
                          assumeAllRequired: Boolean = false,
                          j7: Boolean = false,
                          language: String = "java",
-                         reportFullPath: Boolean = false,
                          useBackticks: Boolean = false,
                          genGetters: Boolean = false,
                          useOptionals: Boolean = false,
@@ -102,9 +100,6 @@ object Main {
         case "--scala" :: rest =>
           traverseList(rest, opts.copy(language = "scala"))
 
-        case "--scala:fp" :: rest =>
-          traverseList(rest, opts.copy(reportFullPath = true))
-
         case "--scala:bt" :: rest =>
           traverseList(rest, opts.copy(useBackticks = true))
 
@@ -129,6 +124,10 @@ object Main {
 
         case "--tpl.cp" :: prefixComment :: rest =>
           templateOpts = templateOpts.copy(commentPrefix = prefixComment)
+          traverseList(rest, opts)
+
+        case "--scala:fp" :: rest =>
+          println( "ignoring obsolete option: --scala:fp")
           traverseList(rest, opts)
 
         case opt :: _ =>
@@ -161,7 +160,6 @@ object Main {
     val genOpts = GenOpts(opts.packageName, opts.className,
                           assumeAllRequired = opts.assumeAllRequired,
                           j7 = opts.j7,
-                          reportFullPath = opts.reportFullPath,
                           useBackticks = opts.useBackticks,
                           genGetters = opts.genGetters,
                           useOptionals = opts.useOptionals,